LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate-expected.txt


FAIL pointerrawupdate event is not fired assert_equals: events received expected "pointermove,pointerdown,pointermove,pointermove,pointerup" but got "pointermove,pointerdown,pointermove,pointermove,pointermove,pointerup,pointermove"
PASS onpointerrawupdate is not exposed
